rename http-server->apiserver, add config to collector.srv
