+++ shortener/internal/apiserver/handlers/url/save/save_test.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"linkify/internal/http-server/handlers/url/save"
-	mocker "linkify/internal/http-server/handlers/url/save/mocks"
+	"linkify/internal/apiserver/handlers/url/save"
+	mocker "linkify/internal/apiserver/handlers/url/save/mocks"
 	"linkify/internal/lib/logger/handlers/slogdiscard"
 	"net/http"
 	"net/http/httptest"
